+++ tests/math/tests_pipeline.cpp
@@ -8,27 +8,27 @@ class PipelineTest : public ::testing::Test
 {
 protected:
   // The VRP (View Reference Point) of the SRC.
-  core::Vector3 vrp = {25, 15, 80};
+  core::Vector3 vrp = {25.0f, 15.0f, 80.0f};
   // The FP (Focal Point) of the SRC.
-  core::Vector3 p = {20, 10, 25};
+  core::Vector3 p = {20.0f, 10.0f, 25.0f};
 
   // The distance from the VRP to the focal point.
-  double dp = 40;
+  float dp = 40;
 
   // The window of the SRC.
-  core::Vector2 min_window = {0, 0};
-  core::Vector2 max_window = {16, 12};
+  core::Vector2 min_window = {0.0f, 0.0f};
+  core::Vector2 max_window = {16.0f, 12.0f};
   // The viewport of the SRC.
-  core::Vector2 min_viewport = {0, 0};
-  core::Vector2 max_viewport = {319, 239};
+  core::Vector2 min_viewport = {0.0f, 0.0f};
+  core::Vector2 max_viewport = {319.0f, 239.0f};
 
   // The point that will be projected.
   // this points form a pyramid having 5 faces.
-  core::Vector4 a = {21.2, 0.7, 42.3, 1};
-  core::Vector4 b = {34.1, 3.4, 27.2, 1};
-  core::Vector4 c = {18.8, 5.6, 14.6, 1};
-  core::Vector4 d = {5.9, 2.9, 29.7, 1};
-  core::Vector4 e = {20, 20, 31.6, 1};
+  core::Vector4 a = {21.2f, 0.7f, 42.3f, 1.0f};
+  core::Vector4 b = {34.1f, 3.4f, 27.2f, 1.0f};
+  core::Vector4 c = {18.8f, 5.6f, 14.6f, 1.0f};
+  core::Vector4 d = {5.9f, 2.9f, 29.7f, 1.0f};
+  core::Vector4 e = {20.0f, 20.0f, 31.6f, 1.0f};
 
   void SetUp() override {}
 };
@@ -42,17 +42,17 @@ TEST_F(PipelineTest, sru2src)
   core::Matrix result = math::sru_to_src(vrp, p);
 
   std::vector<std::vector<float>> expected = {
-      {0.995893206467705, 0.000000000000000, -0.090535746042518, -17.654470478291100},
-      {-0.008163333305850, 0.995926663313714, -0.089796666364351, -7.551083307911350},
-      {0.090166963466743, 0.090166963466743, 0.991836598134176, -82.953606389403800},
-      {0.000000000000000, 0.000000000000000, 0.000000000000000, 1.000000000000000}};
+      {0.995893206467705f, 0.000000000000000f, -0.090535746042518f, -17.654470478291100f},
+      {-0.008163333305850f, 0.995926663313714f, -0.089796666364351f, -7.551083307911350f},
+      {0.090166963466743f, 0.090166963466743f, 0.991836598134176f, -82.953606389403800f},
+      {0.000000000000000f, 0.000000000000000f, 0.000000000000000f, 1.000000000000000f}};
 
   for (int i = 0; i < 4; i++)
   {
     for (int j = 0; j < 4; j++)
     {
       // after the 3rd decimal place, the error is insignificant.
-      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001);
+      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001f);
     }
   }
 };
@@ -66,17 +66,17 @@ TEST_F(PipelineTest, projection)
   core::Matrix result = math::projection(vrp, p, dp);
 
   std::vector<std::vector<float>> expected = {
-      {1, 0, 0, 0},
-      {0, 1, 0, 0},
-      {0, 0, 1, 0},
-      {0, 0, -0.025, 0}};
+      {1.0f, 0.0f, 0.0f, 0.0f},
+      {0.0f, 1.0f, 0.0f, 0.0f},
+      {0.0f, 0.0f, 1.0f, 0.0f},
+      {0.0f, 0.0f, -0.025f, 0.0f}};
 
   for (int i = 0; i < 4; i++)
   {
     for (int j = 0; j < 4; j++)
     {
       // after the 3rd decimal place, the error is insignificant.
-      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001);
+      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001f);
     }
   }
 }
@@ -90,17 +90,17 @@ TEST_F(PipelineTest, src2srt_with_reflection)
   core::Matrix result = math::src_to_srt(min_window, min_viewport, max_window, max_viewport, true);
 
   std::vector<std::vector<float>> expected = {
-      {19.9375, 0, 0, 0},
-      {0, -19.9166666666667, 0, 239},
-      {0, 0, 1, 0},
-      {0, 0, 0, 1}};
+      {19.9375f, 0.0f, 0.0f, 0.0f},
+      {0.0f, -19.9166666666667f, 0.0f, 239.0f},
+      {0.0f, 0.0f, 1.0f, 0.0f},
+      {0.0f, 0.0f, 0.0f, 1.0f}};
 
   for (int i = 0; i < 4; i++)
   {
     for (int j = 0; j < 4; j++)
     {
       // after the 3rd decimal place, the error is insignificant.
-      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001);
+      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001f);
     }
   }
 }
@@ -114,17 +114,17 @@ TEST_F(PipelineTest, src2srt_without_reflection)
   core::Matrix result = math::src_to_srt(min_window, min_viewport, max_window, max_viewport, false);
 
   std::vector<std::vector<float>> expected = {
-      {19.9375, 0, 0, 0},
-      {0, 19.9166666666667, 0, 0},
-      {0, 0, 1, 0},
-      {0, 0, 0, 1}};
+      {19.9375f, 0.0f, 0.0f, 0.0f},
+      {0.0f, 19.9166666666667f, 0.0f, 0.0f},
+      {0.0f, 0.0f, 1.0f, 0.0f},
+      {0.0f, 0.0f, 0.0f, 1.0f}};
 
   for (int i = 0; i < 4; i++)
   {
     for (int j = 0; j < 4; j++)
     {
       // after the 3rd decimal place, the error is insignificant.
-      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001);
+      EXPECT_NEAR(result(i, j), expected[i][j], 0.0001f);
     }
   }
 }
@@ -141,10 +141,10 @@ TEST_F(PipelineTest, full_pipeline)
 
   // Each line represent a result vertex.
   std::vector<std::vector<float>> expected = {
-      {-7.40073, 275.99312, -5.05415, -288.44801, -11.91337},
-      {448.77577, 451.39309, 464.46087, 461.84355, 100.27760},
-      {-39.02426, -52.59438, -66.27271, -52.70259, -48.00489},
-      {0.97560, 1.31485, 1.65681, 1.31756, 1.20012}};
+      {-7.40073f, 275.99312f, -5.05415f, -288.44801f, -11.91337f},
+      {448.77577f, 451.39309f, 464.46087f, 461.84355f, 100.27760f},
+      {-39.02426f, -52.59438f, -66.27271f, -52.70259f, -48.00489f},
+      {0.97560f, 1.31485f, 1.65681f, 1.31756f, 1.20012f}};
 
   core::Vector4 result = {0};
   int j = 0;
@@ -153,10 +153,10 @@ TEST_F(PipelineTest, full_pipeline)
   {
     // after the 3rd decimal place, the error is insignificant.
     result = math::MatrixMultiplyVector(result_matrix, vertex);
-    EXPECT_NEAR(result.x, expected[0][j], 0.0001);
-    EXPECT_NEAR(result.y, expected[1][j], 0.0001);
-    EXPECT_NEAR(result.z, expected[2][j], 0.0001);
-    EXPECT_NEAR(result.w, expected[3][j], 0.0001);
+    EXPECT_NEAR(result.x, expected[0][j], 0.0001f);
+    EXPECT_NEAR(result.y, expected[1][j], 0.0001f);
+    EXPECT_NEAR(result.z, expected[2][j], 0.0001f);
+    EXPECT_NEAR(result.w, expected[3][j], 0.0001f);
 
     j++;
   }
